Start, quit, restart extern controllers from the GUI

+++ src/webots/control/WbController.cpp
@@ -123,6 +123,7 @@ WbController::~WbController() {
   // that this function is the last one to be called
   // exception: don't disconnect readyReadStandard*()
   // signals in order to see the latest log messages
+  mRobot->setControllerStarted(false);
   mProcess->disconnect(SIGNAL(finished(int, QProcess::ExitStatus)));
   mProcess->disconnect(SIGNAL(error(QProcess::ProcessError)));
 
@@ -152,7 +153,7 @@ WbController::~WbController() {
         mSocket->flush();  // otherwise the temination packet is not sent
       }
       // kill the process
-      if (!mProcess->waitForFinished(1000)) {
+      if (mProcess->state() != QProcess::NotRunning && !mProcess->waitForFinished(1000)) {
         WbLog::warning(tr("%1: Forced termination (because process didn't terminate itself after 1 second).").arg(name()));
 #ifdef _WIN32
         // on Windows, we need to kill the process as it may not handle the WM_CLOSE message sent by terminate()
@@ -162,7 +163,7 @@ WbController::~WbController() {
 #endif
       }
     }
-  } else if (mProcess)
+  } else if (mProcess && mProcess->state() != QProcess::NotRunning)
     mProcess->terminate();
 
   delete mSocket;
@@ -225,7 +226,7 @@ void WbController::start() {
   if (mCommandLine.isEmpty())  // python has wrong version or Matlab 64 not available
     return;
 
-  info(tr("Starting controller: \"%1\"").arg(mCommandLine));
+  info(tr("Starting controller: %1").arg(mCommandLine));
 
 #ifdef __linux__
   if (!qgetenv("WEBOTS_FIREJAIL_CONTROLLERS").isEmpty() && mRobot->findField("controller")) {
